add userId to freemode request

+++ front/src/services/freemode.service.ts
@@ -1,7 +1,9 @@
 import api from './api'
-import { ref } from 'vue'
+import { computed, ref } from 'vue'
 import { handleApiError } from './error.handler'
 import type { FreeModeBoard } from '@/types/freemode'
+import { useUserInformations } from '@/stores/userInformations.store'
+const userStore = useUserInformations()
 
 export const freeModeBoard = ref()
 export const freeModeBoardLoading = ref(false)
@@ -9,14 +11,17 @@ export const isCreateFreeModeBoardLoading = ref(false)
 export const isError = ref(false)
 export const isUpdateFreeModeBoardLoading = ref(false)
 export const isDeleteFreeModeBoardLoading = ref(false)
+const userId = computed(() => userStore.userInfo.userId)
 
 export const getFreeModeBoardByName = async (ontology: string) => {
   freeModeBoardLoading.value = true
   isError.value = false
 
   try {
+    if (!userId.value) throw new Error('User not loaded')
+
     const { data } = await api.get<FreeModeBoard[]>(
-      `/freemode.php?ontologyName=${encodeURIComponent(ontology)}`,
+      `/freemode.php?ontologyName=${encodeURIComponent(ontology)}&userId=${userId.value}`,
     )
 
     return data.sort((a, b) => a.title.localeCompare(b.title, 'fr', { sensitivity: 'base' }))
@@ -31,11 +36,15 @@ export const getFreeModeBoardByName = async (ontology: string) => {
 export const createFreeModeBoard = async (board: FreeModeBoard) => {
   isCreateFreeModeBoardLoading.value = true
   isError.value = false
+  console.log(userId)
+
+  if (!userId.value) throw new Error('User not loaded')
 
   try {
     const { data } = await api.post('/freemode.php', {
       title: board.title,
       ontologyName: board.ontologyName,
+      userId: userId.value,
       freemodeData: {
         ZoomLevel: board.ZoomLevel,
         Entities: board.Entities,
@@ -56,10 +65,13 @@ export const createFreeModeBoard = async (board: FreeModeBoard) => {
 export const updateFreeModeBoard = async (updateBoard: FreeModeBoard) => {
   isUpdateFreeModeBoardLoading.value = true
 
+  if (!userId.value) throw new Error('User not loaded')
+
   try {
     const { data } = await api.put('/freemode.php', {
       title: updateBoard.title,
       ontologyName: updateBoard.ontologyName,
+      userId: userId.value,
       freemodeData: {
         ZoomLevel: updateBoard.ZoomLevel,
         Entities: updateBoard.Entities,
@@ -80,9 +92,11 @@ export const deleteFreeModeBoardById = async (freemodeId: string) => {
   isDeleteFreeModeBoardLoading.value = true
   isError.value = false
 
+  if (!userId.value) throw new Error('User not loaded')
+
   try {
     const { data } = await api.delete('/freemode.php', {
-      data: { id: freemodeId },
+      data: { id: freemodeId, userId: userId.value },
     })
 
     return data
@@ -90,6 +104,6 @@ export const deleteFreeModeBoardById = async (freemodeId: string) => {
     if (error) isError.value = true
     handleApiError(error)
   } finally {
-    freeModeBoardLoading.value = false
+    isDeleteFreeModeBoardLoading.value = false
   }
 }
